Split out rendering.
