+++ ksp_plugin/interface.cpp
@@ -229,13 +229,6 @@ QP principia__CelestialWorldDegreesOfFreedom(Plugin const* const plugin,
       ToQP(plugin->CelestialWorldDegreesOfFreedom(index, part_at_origin)));
 }
 
-void principia__ClearTargetVessel(Plugin* const plugin) {
-  journal::Method<journal::ClearTargetVessel> m({plugin});
-  CHECK_NOTNULL(plugin);
-  plugin->ClearTargetVessel();
-  return m.Return();
-}
-
 double principia__CurrentTime(Plugin const* const plugin) {
   journal::Method<journal::CurrentTime> m({plugin});
   CHECK_NOTNULL(plugin);
@@ -368,13 +361,6 @@ QP principia__GetPartActualDegreesOfFreedom(Plugin const* const plugin,
   CHECK_NOTNULL(plugin);
   return m.Return(
       ToQP(plugin->GetPartActualDegreesOfFreedom(part_id, part_at_origin)));
-}
-
-// Returns the frame last set by |plugin->SetPlottingFrame|.  No transfer of
-// ownership.  The returned pointer is never null.
-NavigationFrame const* principia__GetPlottingFrame(Plugin const* const plugin) {
-  journal::Method<journal::GetPlottingFrame> m({plugin});
-  return m.Return(CHECK_NOTNULL(plugin)->GetPlottingFrame());
 }
 
 int principia__GetStderrLogging() {
@@ -711,116 +697,6 @@ void principia__PrepareToReportCollisions(Plugin* const plugin) {
   return m.Return();
 }
 
-Iterator* principia__RenderedPrediction(Plugin* const plugin,
-                                        char const* const vessel_guid,
-                                        XYZ const sun_world_position) {
-  journal::Method<journal::RenderedPrediction> m({plugin,
-                                                  vessel_guid,
-                                                  sun_world_position});
-  CHECK_NOTNULL(plugin);
-  auto const& prediction = plugin->GetVessel(vessel_guid)->prediction();
-  auto rendered_trajectory = plugin->RenderBarycentricTrajectoryInWorld(
-                                 prediction.Begin(),
-                                 prediction.End(),
-                                 FromXYZ<Position<World>>(sun_world_position));
-  return m.Return(new TypedIterator<DiscreteTrajectory<World>>(
-      std::move(rendered_trajectory),
-      plugin));
-}
-
-void principia__RenderedPredictionApsides(Plugin const* const plugin,
-                                          char const* const vessel_guid,
-                                          int const celestial_index,
-                                          XYZ const sun_world_position,
-                                          Iterator** const apoapsides,
-                                          Iterator** const periapsides) {
-  journal::Method<journal::RenderedPredictionApsides> m(
-      {plugin, vessel_guid, celestial_index, sun_world_position},
-      {apoapsides, periapsides});
-  CHECK_NOTNULL(plugin);
-  auto const& prediction = plugin->GetVessel(vessel_guid)->prediction();
-  std::unique_ptr<DiscreteTrajectory<World>> rendered_apoapsides;
-  std::unique_ptr<DiscreteTrajectory<World>> rendered_periapsides;
-  plugin->ComputeAndRenderApsides(celestial_index,
-                                  prediction.Begin(),
-                                  prediction.End(),
-                                  FromXYZ<Position<World>>(sun_world_position),
-                                  rendered_apoapsides,
-                                  rendered_periapsides);
-  *apoapsides = new TypedIterator<DiscreteTrajectory<World>>(
-      check_not_null(std::move(rendered_apoapsides)),
-      plugin);
-  *periapsides = new TypedIterator<DiscreteTrajectory<World>>(
-      check_not_null(std::move(rendered_periapsides)),
-      plugin);
-  return m.Return();
-}
-
-void principia__RenderedPredictionClosestApproaches(
-    Plugin const* const plugin,
-    char const* const vessel_guid,
-    XYZ const sun_world_position,
-    Iterator** const closest_approaches) {
-  journal::Method<journal::RenderedPredictionClosestApproaches> m(
-      {plugin, vessel_guid, sun_world_position},
-      {closest_approaches});
-  CHECK_NOTNULL(plugin);
-  auto const& prediction = plugin->GetVessel(vessel_guid)->prediction();
-  std::unique_ptr<DiscreteTrajectory<World>> rendered_closest_approaches;
-  plugin->ComputeAndRenderClosestApproaches(
-      prediction.Begin(),
-      prediction.End(),
-      FromXYZ<Position<World>>(sun_world_position),
-      rendered_closest_approaches);
-  *closest_approaches = new TypedIterator<DiscreteTrajectory<World>>(
-      check_not_null(std::move(rendered_closest_approaches)),
-      plugin);
-  return m.Return();
-}
-
-void principia__RenderedPredictionNodes(Plugin const* const plugin,
-                                        char const* const vessel_guid,
-                                        XYZ const sun_world_position,
-                                        Iterator** const ascending,
-                                        Iterator** const descending) {
-  journal::Method<journal::RenderedPredictionNodes> m(
-      {plugin, vessel_guid, sun_world_position},
-      {ascending, descending});
-  CHECK_NOTNULL(plugin);
-  auto const& prediction = plugin->GetVessel(vessel_guid)->prediction();
-  std::unique_ptr<DiscreteTrajectory<World>> rendered_ascending;
-  std::unique_ptr<DiscreteTrajectory<World>> rendered_descending;
-  plugin->ComputeAndRenderNodes(prediction.Begin(),
-                                prediction.End(),
-                                FromXYZ<Position<World>>(sun_world_position),
-                                rendered_ascending,
-                                rendered_descending);
-  *ascending = new TypedIterator<DiscreteTrajectory<World>>(
-      check_not_null(std::move(rendered_ascending)),
-      plugin);
-  *descending = new TypedIterator<DiscreteTrajectory<World>>(
-      check_not_null(std::move(rendered_descending)),
-      plugin);
-  return m.Return();
-}
-
-Iterator* principia__RenderedVesselTrajectory(Plugin const* const plugin,
-                                              char const* const vessel_guid,
-                                              XYZ const sun_world_position) {
-  journal::Method<journal::RenderedVesselTrajectory> m({plugin,
-                                                        vessel_guid,
-                                                        sun_world_position});
-  CHECK_NOTNULL(plugin);
-  auto const& psychohistory = plugin->GetVessel(vessel_guid)->psychohistory();
-  auto rendered_trajectory = plugin->RenderBarycentricTrajectoryInWorld(
-                                 psychohistory.Begin(),
-                                 psychohistory.End(),
-                                 FromXYZ<Position<World>>(sun_world_position));
-  return m.Return(new TypedIterator<DiscreteTrajectory<World>>(
-      std::move(rendered_trajectory),
-      plugin));
-}
-
 void principia__ReportCollision(Plugin const* const plugin,
                                 PartId const part1_id,
                                 PartId const part2_id) {
@@ -909,33 +785,11 @@ void principia__SetPartApparentDegreesOfFreedom(Plugin* const plugin,
   return m.Return();
 }
 
-// |navigation_frame| must not be null.  No transfer of ownership of
-// |*navigation_frame|, takes ownership of |**navigation_frame|, nulls
-// |*navigation_frame|.
-void principia__SetPlottingFrame(Plugin* const plugin,
-                                 NavigationFrame** const navigation_frame) {
-  journal::Method<journal::SetPlottingFrame> m({plugin, navigation_frame},
-                                               {navigation_frame});
-  CHECK_NOTNULL(plugin);
-  plugin->SetPlottingFrame(TakeOwnership(navigation_frame));
-  return m.Return();
-}
-
 void principia__SetPredictionLength(Plugin* const plugin,
                                     double const t) {
   journal::Method<journal::SetPredictionLength> m({plugin, t});
   CHECK_NOTNULL(plugin);
   plugin->SetPredictionLength(t * Second);
-  return m.Return();
-}
-
-void principia__SetTargetVessel(Plugin* const plugin,
-                                char const* const vessel_guid,
-                                int const reference_body_index) {
-  journal::Method<journal::SetTargetVessel> m(
-      {plugin, vessel_guid, reference_body_index});
-  CHECK_NOTNULL(plugin);
-  plugin->SetTargetVessel(vessel_guid, reference_body_index);
   return m.Return();
 }
