+++ ksp_plugin_test/ksp_plugin_test.vcxproj
@@ -39,6 +39,7 @@
     <ClCompile Include="..\ksp_plugin\interface.cpp" />
     <ClCompile Include="..\ksp_plugin\interface_flight_plan.cpp" />
     <ClCompile Include="..\ksp_plugin\interface_iterator.cpp" />
+    <ClCompile Include="..\ksp_plugin\interface_renderer.cpp" />
     <ClCompile Include="..\ksp_plugin\interface_vessel.cpp" />
     <ClCompile Include="..\ksp_plugin\part.cpp" />
     <ClCompile Include="..\ksp_plugin\part_subsets.cpp" />
@@ -47,6 +48,7 @@
     <ClCompile Include="..\ksp_plugin\vessel.cpp" />
     <ClCompile Include="celestial_test.cpp" />
     <ClCompile Include="flight_plan_test.cpp" />
+    <ClCompile Include="interface_renderer_test.cpp" />
     <ClCompile Include="interface_test.cpp" />
     <ClCompile Include="manœuvre_test.cpp" />
     <ClCompile Include="mock_flight_plan.cpp" />
