+++ testing_utilities/testing_utilities.vcxproj
@@ -27,6 +27,7 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="actions.hpp" />
     <ClInclude Include="algebra.hpp" />
     <ClInclude Include="algebra_body.hpp" />
     <ClInclude Include="almost_equals.hpp" />
