+++ testing_utilities/testing_utilities.vcxproj.filters
@@ -71,6 +71,9 @@
     <ClInclude Include="matchers.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="actions.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="almost_equals_test.cpp">
